requirements.txt
streamlit==1.41.1
chromadb==0.5.23
langchain-ollama==0.2.2
langgraph==0.2.60
langchain-core==0.3.28
langchain-groq==0.2.2
python-dotenv==1.0.1